@@ -1,7 +1,7 @@
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 import './styles/index.css';
 
 const container = document.getElementById('root');
 
-ReactDOM.render(<App />, container);
+createRoot(container).render(<App />);
