@@ -13,7 +13,7 @@ module.exports = {
 		'plugin:react/jsx-runtime',
 		'plugin:react-hooks/recommended',
 		'standard',
-		'eslint-config-prettier'
+		'prettier'
 	],
 	parserOptions: {
 		ecmaFeatures: {
